% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{leading_na}
\alias{leading_na}
\title{Leading NA}
\usage{
leading_na(x)
}
\arguments{
\item{x}{An integer vector}
}
\description{
This function returns a logical vector identifying if
there are leading NA, marking the leadings NA as TRUE and
everything else as FALSE.
}
