\name{privateMST-package}
\alias{privateMST-package}
\alias{privateMST}
\docType{package}
\title{
 Two private minimum spanning tree algorithms
}
\description{
  This package provides two methods to compute an almost minimum spanning tree under differential privacy constraints.
}
\details{
  The two main functions offered by the package are laplaceMST and PrivateSpanningTree. They respectively compute spanning trees according to the Laplace and PAMST/Kruskal methods.
}
\author{
Your Name, email optional.

Maintainer: Your Name <your@email.com>
}
\keyword{ Minimum spanning tree, differential privacy, Laplace mechanism, PAMST algorithm, Private Kruskal algorithm}
