% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{privateMST}
\alias{privateMST}
\alias{privateMST-package}
\title{privateMST}
\description{
This package offers two functions for releasing a minimum spanning tree under privacy constraints
}
\author{
you
}
